+++ scripts/hooks.js
@@ -16,29 +16,29 @@ Hooks.on('renderTokenHUD', (tokenHUD, html, app) => {
     const actor = token?.actor;
     const engine = game.stealthy.engine;
 
-    const hidden = actor?.effects.find(e => e.label === game.i18n.localize("stealthy-hidden-label") && !e.disabled);
-    if (hidden) {
-      let { flag, value } = engine.getHiddenFlagAndValue(hidden);
+    const hiddenEffect = engine.findHiddenEffect(actor);
+    if (hiddenEffect) {
+      let { flag, value } = engine.getHiddenFlagAndValue(actor, hiddenEffect);
       const inputBox = $(
-        `<input id="ste_hid_inp_box" title="${game.i18n.localize("stealthy-hidden-inputBox-title")}" type="text" name="hidden_value_inp_box" value="${value}"></input>`
+        `<input id="ste_hid_inp_box" title="${game.i18n.localize("stealthy.hidden.inputBox")}" type="text" name="hidden_value_inp_box" value="${value}"></input>`
       );
       html.find(".right").append(inputBox);
       inputBox.change(async (inputbox) => {
         if (token === undefined) return;
-        await engine.setHiddenValue(actor, duplicate(hidden), flag, Number(inputbox.target.value));
+        await engine.setHiddenValue(actor, duplicate(hiddenEffect), flag, Number(inputbox.target.value));
       });
     }
 
-    const spot = actor?.effects.find(e => e.label === game.i18n.localize("stealthy-spot-label") && !e.disabled);
-    if (spot) {
-      let { flag, value } = engine.getSpotFlagAndValue(spot);
+    const spotEffect = engine.findSpotEffect(actor);
+    if (spotEffect) {
+      let { flag, value } = engine.getSpotFlagAndValue(actor, spotEffect);
       const inputBox = $(
-        `<input id="ste_spt_inp_box" title="${game.i18n.localize("stealthy-spot-inputBox-title")}" type="text" name="spot_value_inp_box" value="${value}"></input>`
+        `<input id="ste_spt_inp_box" title="${game.i18n.localize("stealthy.spot.inputBox")}" type="text" name="spot_value_inp_box" value="${value}"></input>`
       );
       html.find(".left").append(inputBox);
       inputBox.change(async (inputbox) => {
         if (token === undefined) return;
-        await engine.setSpotValue(actor, duplicate(spot), flag, Number(inputbox.target.value));
+        await engine.setSpotValue(actor, duplicate(spotEffect), flag, Number(inputbox.target.value));
       });
     }
   }
@@ -50,7 +50,7 @@ Hooks.on('getSceneControlButtons', (controls) => {
   tokenControls.tools.push({
     icon: 'fa-solid fa-eyes',
     name: 'stealthy-spotting',
-    title: game.i18n.localize("stealthy-active-spot"),
+    title: game.i18n.localize("stealthy.activeSpot"),
     toggle: true,
     active: game.stealthy.activeSpot,
     onClick: (toggled) => game.stealthy.socket.executeForEveryone('ToggleActiveSpot', toggled)
@@ -58,6 +58,8 @@ Hooks.on('getSceneControlButtons', (controls) => {
 });
 
 Hooks.once('ready', async () => {
+  if (!game.modules.get('lib-wrapper')?.active && game.user.isGM)
+    ui.notifications.error("Stealthy requires the 'libWrapper' module. Please install and activate it.");
   if (!game.user.isGM)
     game.stealthy.activeSpot = await game.stealthy.socket.executeAsGM('GetActiveSpot');
 });