+++ scripts/patches.js
@@ -2,9 +2,9 @@ import { Stealthy } from "./stealthy.js";
 
 Hooks.once('setup', () => {
   libWrapper.register(
-    'stealthy',
+    Stealthy.MODULE_ID,
     'DetectionModeBasicSight.prototype.testVisibility',
-    (wrapped, visionSource, mode, config = {}) => {
+    function (wrapped, visionSource, mode, config = {}) {
       const engine = game.stealthy.engine;
       if (!engine.testStealth(visionSource, config)) return false;
       return engine.basicVision(wrapped, visionSource, mode, config);
@@ -14,9 +14,9 @@ Hooks.once('setup', () => {
   );
 
   libWrapper.register(
-    'stealthy',
+    Stealthy.MODULE_ID,
     'DetectionModeInvisibility.prototype.testVisibility',
-    (wrapped, visionSource, mode, config = {}) => {
+    function (wrapped, visionSource, mode, config = {}) {
       const engine = game.stealthy.engine;
       if (!engine.testStealth(visionSource, config)) return false;
       return engine.seeInvisibility(wrapped, visionSource, mode, config);